Split main.c file and remove buggy parsing code
Splits the contents of the main.c file into smaller files for each of
the relevant stucts and its functions.

Removes the existing model parsing code. It was hastily written and
buggy so I will rewrite it later.

Allow for the Makefile to wildcard all the new sourcefiles, to be able
to watch for changes and correctly compile. Using one step compile, no
intermediate object with linking.