Basic wireframe model rendering
Fixes implementation of Bresenham's Line algorithm. The existing
implementation was manually derived by me, but was buggy and prevented
the proper rendering of the model.

Functions to parse vec3f and face structs from an input line now return
struct by value instead of populating pointers to numeric varibles.

Functions to find the minimum and maximum value of a vec3f struct
now return by value instead of populating pointers to floats.

Adds simple code to the main function to load the model, fetch vertexes
for each triangle face and draw the lines of the triangle.